+++ h files/Citizen.h
@@ -21,7 +21,7 @@ class Citizen {
         virtual string getWorkplace();
         virtual float getSatisfaction();
         virtual void displayInfo() = 0;
-        void changeSatisfaction(int amt);
+        void changeSatisfaction(float amt);
 };
 
 // Concrete Products
